+++ src/components/LanguageSwitcher/LanguageSwitcher.module.scss
@@ -2,6 +2,9 @@
 
 .switcher_container {
   margin: 0 30px;
+  @media (max-width: $mobile) {
+    margin: 0 10px;
+  }
 }
 .switcher {
   cursor: pointer;
@@ -9,7 +12,6 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 100%;
   border-radius: 10px;
   border: 0;
   background-color: $color-submenu-background;
@@ -19,5 +21,16 @@
     padding: 10px;
     border-radius: 10px;
     color: white;
+  }
+  @media (max-width: $mobile) {
+    width: 100%;
+    aspect-ratio: 1;
+    padding: 0;
+    height: 16px;
+    .switcher_item {
+      padding: 5px;
+      height: 16px;
+      font-size: 10px;
+    }
   }
 }
